Part 3: Authentication and Cross Site Origin Requests/requirements/base.txt
django==1.11.6
django-environ==0.4.4
djangorestframework==3.7.0